+++ src/pages/login/Login.tsx
@@ -5,7 +5,6 @@ import Password from '@mui/icons-material/LockOpen'
 import { Button } from "../../components/common";
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Troubleshoot } from "@mui/icons-material";
 
 
 export const Login  = () => {
@@ -14,29 +13,7 @@ export const Login  = () => {
     const navigate = useNavigate();
     const [isFirst, setIsFirst] = useState(false); 
      
-    const fetchData = async () => {
-      try {
-        const response = await fetch(`http://aniroomi-env.eba-rj7upyms.ap-northeast-2.elasticbeanstalk.com/style`, {
-          method: 'GET',
-          credentials: 'include',
-        });
-  
-        if (response.ok) {
-          const data = await response.json();
-          setIsFirst(data);
-          if(isFirst === false) {
-              navigate('/testpage');
-            }
-            else{
-              navigate('/home');
-            }
-        } else {
-          console.error('Failed to fetch initial card data: ', response.status, response.statusText);
-        }
-      } catch (error) {
-        console.error('Failed to fetch initial card data: ', error);
-      }
-    };
+
     const handleSubmit = async (e: React.FormEvent) => {
         e.preventDefault();
         console.log("sdfs")
@@ -56,13 +33,9 @@ export const Login  = () => {
 
           // Check if login is successful, then redirect to StarPage
           if (response.ok) {
-            // if(!isFirst) {
-            //   navigate('/testpage');
-            // }
-            // else{
-            //   navigate('/home');
-            // }
-            fetchData();
+           
+            navigate('/home');
+            
             
               }
         } catch (error) {
